+++ tests/WebHemiTest/Auth/AuthTest.php
@@ -25,9 +25,9 @@ use WebHemi\Data\Storage\User\UserStorage;
 use WebHemiTest\Fixtures\EmptyAuthStorage;
 use WebHemiTest\AssertTrait;
 use WebHemiTest\InvokePrivateMethodTrait;
-use PHPUnit_Framework_TestCase as TestCase;
-use PHPUnit_Framework_IncompleteTestError;
-use PHPUnit_Framework_SkippedTestError;
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\IncompleteTestError;
+use PHPUnit\Framework\SkippedTestError;
 
 /**
  * Class AuthTest
@@ -53,7 +53,7 @@ class AuthTest extends TestCase
     protected function checkRequirements()
     {
         if (!extension_loaded('pdo_sqlite')) {
-            throw new PHPUnit_Framework_SkippedTestError('No SQLite Available');
+            throw new SkippedTestError('No SQLite Available');
         }
 
         parent::checkRequirements();
@@ -79,7 +79,7 @@ class AuthTest extends TestCase
                 $result = self::$dataDriver->query($sql);
 
                 if (!$result) {
-                    throw new PHPUnit_Framework_IncompleteTestError(
+                    throw new IncompleteTestError(
                         'Cannot set up test database: '.json_encode(self::$dataDriver->errorInfo()).'; query: '.$sql
                     );
                 }
@@ -94,7 +94,7 @@ class AuthTest extends TestCase
         $result = $statement->execute();
 
         if (!$result) {
-            throw new PHPUnit_Framework_IncompleteTestError(
+            throw new IncompleteTestError(
                 'Cannot set up test database: '.json_encode(self::$dataDriver->errorInfo()).'; query: '.$sql
             );
         }