+++ tests/WebHemiTest/Routing/ResultTest.php
@@ -14,7 +14,7 @@ namespace WebHemiTest\Routing;
 use InvalidArgumentException;
 use WebHemi\Router\Result;
 use WebHemiTest\AssertTrait;
-use PHPUnit_Framework_TestCase as TestCase;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Class ResultTest.